Export main function of `znet`
